+++ src/examples/pong/aiSystem.h
@@ -11,5 +11,5 @@ class AiSystem : public System
   AiSystem();
   ~AiSystem();
 
-  void Update(double dt, Registry* registry, Input* input);
+  void Update(float dt, Registry* registry, Input* input);
 };
