+++ src/examples/pong/playerSystem.cc
@@ -8,7 +8,7 @@ PlayerSystem::~PlayerSystem()
 {
 }
 
-void PlayerSystem::Update(double dt, Registry* registry, Input* input)
+void PlayerSystem::Update(float dt, Registry* registry, Input* input)
 {
   PongRules& pongRules = registry->GetComponent<PongRules>();
   RigidBody* playerRigidBody = registry->GetComponent<RigidBody>(pongRules.playerPaddleID);
