+++ src/systems/audioSystem.h
@@ -14,5 +14,5 @@ class AudioSystem : public System
   AudioSystem();
   ~AudioSystem();
 
-  void Update(double dt, Registry* registry, Input* input);
+  void Update(float dt, Registry* registry, Input* input);
 };
